separar models en directorio y agregar crud clinics
