+++ app/api/dependencies.py
@@ -3,7 +3,7 @@ from fastapi.security import OAuth2PasswordBearer
 from sqlalchemy.orm import Session
 from app.core.database import get_db
 from app.services.auth_service import get_current_user_from_token
-from app.domain import models
+from app.domain.models import User
 
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/api/v1/auth/token")
 
@@ -14,7 +14,7 @@ def get_current_user(token: str = Depends(oauth2_scheme), db: Session = Depends(
     return user
 
 def has_role(required_role_name: str):
-    def role_checker(current_user: models.User = Depends(get_current_user)):
+    def role_checker(current_user: User = Depends(get_current_user)):
         if not current_user.role or current_user.role.name != required_role_name:
             raise HTTPException(
                 status_code=status.HTTP_403_FORBIDDEN,
@@ -25,7 +25,7 @@ def has_role(required_role_name: str):
 
 def require_permission(permission_name: str):
     async def _require_permission(
-        current_user: models.User = Depends(get_current_user),
+        current_user: User = Depends(get_current_user),
         db: Session = Depends(get_db)
     ):
         if not current_user.role:
@@ -49,11 +49,11 @@ def require_permission(permission_name: str):
     return _require_permission
 
 # Ejemplos de roles específicos
-def is_admin(current_user: models.User = Depends(get_current_user)):
+def is_admin(current_user: User = Depends(get_current_user)):
     return has_role("admin")(current_user)
 
-def is_doctor(current_user: models.User = Depends(get_current_user)):
+def is_doctor(current_user: User = Depends(get_current_user)):
     return has_role("doctor")(current_user)
 
-def is_patient(current_user: models.User = Depends(get_current_user)):
+def is_patient(current_user: User = Depends(get_current_user)):
     return has_role("patient")(current_user)
