+++ app/data/repositories/audit_log_repository.py
@@ -1,25 +1,27 @@
 from sqlalchemy.orm import Session
-from app.domain import models, schemas
+from app.domain import schemas
 from typing import List, Optional
+
+from app.domain.models import AuditLog
 
 class AuditLogRepository:
     def __init__(self, db: Session):
         self.db = db
 
-    def create(self, audit_log: schemas.AuditLogCreate) -> models.AuditLog:
-        db_audit_log = models.AuditLog(**audit_log.model_dump())
+    def create(self, audit_log: schemas.AuditLogCreate) -> AuditLog:
+        db_audit_log = AuditLog(**audit_log.model_dump())
         self.db.add(db_audit_log)
         self.db.commit()
         self.db.refresh(db_audit_log)
         return db_audit_log
 
-    def get_by_id(self, audit_log_id: int) -> Optional[models.AuditLog]:
-        return self.db.query(models.AuditLog).filter(models.AuditLog.id == audit_log_id).first()
+    def get_by_id(self, audit_log_id: int) -> Optional[AuditLog]:
+        return self.db.query(AuditLog).filter(AuditLog.id == audit_log_id).first()
 
-    def get_all(self, skip: int = 0, limit: int = 100) -> List[models.AuditLog]:
-        return self.db.query(models.AuditLog).offset(skip).limit(limit).all()
+    def get_all(self, skip: int = 0, limit: int = 100) -> List[AuditLog]:
+        return self.db.query(AuditLog).offset(skip).limit(limit).all()
 
-    def update(self, audit_log_id: int, audit_log: schemas.AuditLogUpdate) -> Optional[models.AuditLog]:
+    def update(self, audit_log_id: int, audit_log: schemas.AuditLogUpdate) -> Optional[AuditLog]:
         db_audit_log = self.get_by_id(audit_log_id)
         if db_audit_log:
             update_data = audit_log.model_dump(exclude_unset=True)
