+++ app/domain/models/__init__.py
@@ -1,0 +1,30 @@
+from .appointment import Appointment
+from .clinic import Clinic
+from .permission import Permission
+from .rolepermission import RolePermission
+from .role import Role
+from .auditlog import AuditLog
+from .clinicalprotocol import ClinicalProtocol
+from .clinicalstudy import ClinicalStudy
+from .resource import Resource
+from .service import Service
+from .appointmentservice import AppointmentService
+from .patient import Patient
+from .user import User
+
+
+__all__ = [
+    "Appointment",
+    "Clinic",
+    "Permission",
+    "RolePermission",
+    "Role",
+    "AuditLog",
+    "ClinicalProtocol",
+    "ClinicalStudy",
+    "Resource",
+    "Service",
+    "AppointmentService",
+    "Patient",
+    "User"
+]
