+++ app/domain/models/auditlog.py
@@ -1,0 +1,24 @@
+from sqlalchemy import Column, Integer, String, Boolean, DateTime, ForeignKey, Enum, Text
+from sqlalchemy.orm import relationship, Mapped, mapped_column
+from sqlalchemy.sql import func
+from app.core.database import Base
+from sqlalchemy.dialects.mysql import JSON
+from typing import List, Optional
+
+class AuditLog(Base):
+    __tablename__ = "audit_logs"
+
+    id = Column(Integer, primary_key=True, index=True)
+    user_id = Column(Integer, ForeignKey("users.id"))
+    action = Column(String(100), nullable=False)
+    entity_type = Column(String(100))
+    entity_id = Column(Integer)
+    old_values = Column(JSON)
+    new_values = Column(JSON)
+    ip_address = Column(String(50))
+    user_agent = Column(String(255))
+    created_at = Column(DateTime, default=func.now(), nullable=False)
+
+    clinic_id = Column(Integer, ForeignKey("clinics.id"))
+    user = relationship("User", foreign_keys=[user_id])
+    clinic = relationship("Clinic", back_populates="audit_logs")
