+++ app/domain/models/permission.py
@@ -1,0 +1,20 @@
+from sqlalchemy import Column, Integer, String, Boolean, DateTime, ForeignKey, Enum, Text
+from sqlalchemy.orm import relationship, Mapped, mapped_column
+from sqlalchemy.sql import func
+from app.core.database import Base
+from sqlalchemy.dialects.mysql import JSON
+from typing import List, Optional
+
+class Permission(Base):
+    __tablename__ = "permissions"
+
+    id = Column(Integer, primary_key=True, index=True)
+    name = Column(String(100), unique=True, nullable=False)
+    description = Column(String(255))
+    created_at = Column(DateTime, default=func.now())
+    updated_at = Column(DateTime, default=func.now(), onupdate=func.now())
+
+    role_permissions = relationship("RolePermission", back_populates="permission")
+
+    def __repr__(self):
+        return f"<Permission(name='{self.name}')>"
