+++ app/domain/models/role.py
@@ -1,0 +1,25 @@
+from sqlalchemy import Column, Integer, String, Boolean, DateTime, ForeignKey, Enum, Text
+from sqlalchemy.orm import relationship, Mapped, mapped_column
+from sqlalchemy.sql import func
+from app.core.database import Base
+from sqlalchemy.dialects.mysql import JSON
+from typing import List, Optional
+
+class Role(Base):
+    __tablename__ = "roles"
+
+    id = Column(Integer, primary_key=True, index=True)
+    name = Column(String(50), unique=True, index=True, nullable=False)
+    description = Column(String(255))
+    created_at = Column(DateTime, default=func.now())
+    updated_at = Column(DateTime, default=func.now(), onupdate=func.now())
+
+    users = relationship("User", back_populates="role")
+    role_permissions = relationship("RolePermission", back_populates="role")
+
+    @property
+    def permissions(self):
+        return [rp.permission for rp in self.role_permissions]
+
+    def __repr__(self):
+        return f"<Role(name='{self.name}')>"
